Refactor infer API
